+++ packages/runtime/src/index.ts
@@ -1,5 +1,7 @@
 export * from "./RuntimeEvent";
+export * from "./RuntimeHostConfiguration";
 export * from "./RuntimeHost";
+export * from "./RuntimeConfigurationValidator";
 export * from "./RuntimeModule";
 export * from "./RuntimeModuleSnapshot";
 export * from "./RuntimeModuleStatus";